Correcao da activity e troca para a detalhes activity
