+++ app/src/main/java/br/com/maonamassa/olharhorizontal/activities/ListaActivity.kt
@@ -58,7 +58,7 @@ class ListaActivity: AppCompatActivity(), OngClickListener {
     }
 
     override fun onOngClicked(ong: ONG) {
-        val proximaTela = Intent(this, MapActivity::class.java)
+        val proximaTela = Intent(this, DetalhesActivity::class.java)
         proximaTela.putExtra("ong", ong)
         startActivity(proximaTela)
     }
