+++ app/src/main/res/layout/activity_map.xml
@@ -28,16 +28,16 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content">
 
-                <ImageView
-                    android:id="@+id/imagemDescricao"
-                    app:layout_collapseMode="parallax"
-                    android:contentDescription="@string/app_name"
-                    android:scaleType="fitXY"
-                    android:background="@color/branco"
-                    android:layout_width="match_parent"
-                    android:layout_height="350dp" />
-
-                <View
+                //<ImageView
+                />/    android:id="@+id/imagemDescricao"
+                //    app:layout_collapseMode="parallax"
+                //    android:contentDescription="@string/app_name"
+                //    android:scaleType="fitXY"
+                //    android:background="@color/branco"
+                //    android:layout_width="match_parent"
+                //    android:layout_height="350dp" />
+//
+                //<View
                     android:layout_width="match_parent"
                     android:layout_height="350dp"
                     android:background="@drawable/card_background"/>
@@ -45,14 +45,14 @@
 
             </RelativeLayout>
 
-            <android.support.v7.widget.Toolbar
-                android:id="@+id/toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
-                app:layout_collapseMode="pin"
-                app:collapseIcon="@drawable/ic_back_button"
-                />
+           //<android.support.v7.widget.Toolbar
+           />/    android:id="@+id/toolbar"
+           //    android:layout_width="match_parent"
+           //    android:layout_height="?attr/actionBarSize"
+           //    app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
+           //    app:layout_collapseMode="pin"
+           //    app:collapseIcon="@drawable/ic_back_button"
+           //    />
 
         </android.support.design.widget.CollapsingToolbarLayout>
 
@@ -87,18 +87,18 @@
                     android:paddingEnd="24dp"
                     android:paddingTop="16dp"/>
 
-                <com.google.android.gms.maps.MapView
-                    android:id="@+id/mapa"
-                    android:layout_width="match_parent"
-                    android:layout_height="200dp"
-                    android:layout_margin="24dp"
-                    map:uiScrollGestures = "true"
-                    map:uiZoomGestures = "true"
-                    map:uiZoomControls = "true"
-                    map:uiRotateGestures = "true"
-                    map:uiTiltGestures = "true"
-
-                    />
+                //<com.google.android.gms.maps.MapView
+                />   android:id="@+id/mapa"
+                //    android:layout_width="match_parent"
+                //    android:layout_height="200dp"
+                //    android:layout_margin="24dp"
+                //    map:uiScrollGestures = "true"
+                //    map:uiZoomGestures = "true"
+                //    map:uiZoomControls = "true"
+                //    map:uiRotateGestures = "true"
+                //    map:uiTiltGestures = "true"
+//
+                //    />
 
                 <TextView
                 android:id="@+id/texto_endereco"
@@ -118,8 +118,9 @@
                     android:layout_marginEnd="24dp"
                     android:layout_marginStart="24dp"
                     android:layout_marginBottom="16dp"
-                    android:text="Site da ONG" />
-                   // android:visibility="gone"
+                    android:text="Site da ONG"
+                    />
+
